refactor: extract PDF front matter splitting
